disable logging for unapproved
