+++ src/i18n/cs.toml
@@ -752,6 +752,7 @@ apiMessageErrorWithMessage = "Chyba při ukládání jízd. Chyba: {error}"
 apiMessageSuccess = "Jízdy byly úspěšně uloženy."
 messageEntryNotEnabled = "Zapisování jízd není v tuto chvíli povoleno."
 messageFormValidationFailed = "Nemáte správně vyplněnou některou jízdu."
+messageUserNotApproved = "Nemůžete zapsat jízdy dokud nejste schválen*a jako člen*ka týmu. Požádejte kolegy o schválení."
 
 [profile]
 buttonDeleteAccount = "Smazat"
@@ -1007,6 +1008,7 @@ buttonSaveChangesCount = "Žádné změny | Uložit {count} změnu | Uložit {co
 buttonTraceMap = "Zakreslit do mapy"
 descriptionRoutes = "<p>Jízdy můžete zapisovat až 8 dní zpětně.</p><p>Pokud kombinujete dopravní prostředky, zapište tu nejdelší část, případně druhou nejdelší je-li na vlastní pohon.</p>"
 hintAutomaticLogging = "Pro automatický záznam můžete využít jednu z těchto aplikací třetích stran.<br />(Ověřit, zda jste propojeni s Cyclers či Na kole Prahou můžete z technických důvodů pouze přímo v aplikacích.)"
+hintLoggingNotApproved = "Nejste schválen*a jako člen*ka týmu. Pro propojení svého účtu s aplikacemi požádejte své kolegy o schválení."
 hintLockedSection = "Tato sekce s vám otevře ihned, jakmile začne výzva."
 hintManualLogging = "Mimo toto webové rozhraní můžete také využívat naše oficiální mobilní aplikace:"
 hintTabApp = "Pro automatický záznam můžete využít jednu z těchto aplikací třetích stran:"
