+++ src/i18n/en.toml
@@ -749,6 +749,7 @@ apiMessageErrorWithMessage = "Error saving trips. Error: {error}"
 apiMessageSuccess = "Trips were successfully saved."
 messageEntryNotEnabled = "Trip logging is not currently allowed."
 messageFormValidationFailed = "Some of your trips are not filled in correctly."
+messageUserNotApproved = "You cannot log rides until you are approved as a team member. Ask your colleagues for approval."
 
 [profile]
 buttonAddEmailField = "Add another email"
@@ -1006,6 +1007,7 @@ buttonSaveChangesCount = "No changes | Save {count} change | Save {count} change
 buttonTraceMap = "Trace route"
 descriptionRoutes = "<p>You can record rides up to 8 days in advance.</p><p>If you combine vehicles, record the longest part, or the second longest if self-driven.</p>"
 hintAutomaticLogging = "You can use one of these third-party apps for automatic route logging.<br />(Due to technical reasons, you can only verify your connection with Cyclers or Na kole Prahou directly in the applications.)"
+hintLoggingNotApproved = "You are not approved as a team member. Ask your colleagues for approval to connect your account with the applications."
 hintLockedSection = "This section will open as soon as the challenge starts."
 hintManualLogging = "In addition to this web interface, you can also use our official mobile apps:"
 hintTabApp = "You can use one of these third-party apps for automatic route logging:"
